requirements.txt
prometheus_client
pyyaml
requests

